add check Role

+++ src/main/java/com/tsu/projectX/controllers/TeamController.java
@@ -2,6 +2,7 @@ package com.tsu.projectX.controllers;
 
 import com.tsu.projectX.data.requestDto.UserTeam;
 import com.tsu.projectX.data.responseDto.TeamResponseDto;
+import com.tsu.projectX.services.interfaces.IAuthService;
 import com.tsu.projectX.services.interfaces.ITeamService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -11,6 +12,8 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 import java.util.UUID;
 
+import static com.tsu.projectX.config.AuthConfig.*;
+
 @RestController
 @CrossOrigin
 @RequestMapping("/teams")
@@ -18,9 +21,18 @@ public class TeamController {
 
     @Autowired
     private ITeamService teamService;
+    @Autowired
+    private IAuthService authService;
 
     @GetMapping("{id}")
-    private ResponseEntity<TeamResponseDto> get(@PathVariable(name = "id") UUID id) {
+    private ResponseEntity<TeamResponseDto> get(
+            @RequestHeader(name = "auth-token") UUID authToken,
+            @PathVariable(name = "id") UUID id) {
+        boolean accessible = authService.checkAuthAndPermission(authToken);
+        if (!accessible) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         TeamResponseDto team = teamService.get(id);
         return team != null
                 ? new ResponseEntity<>(team, HttpStatus.OK)
@@ -28,7 +40,13 @@ public class TeamController {
     }
 
     @GetMapping
-    private ResponseEntity<List<TeamResponseDto>> getAll() {
+    private ResponseEntity<List<TeamResponseDto>> getAll(
+            @RequestHeader(name = "auth-token") UUID authToken) {
+        boolean accessible = authService.checkAuthAndPermission(authToken);
+        if (!accessible) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         List<TeamResponseDto> teams = teamService.getAll();
         return teams != null
                 ? new ResponseEntity<>(teams, HttpStatus.OK)
@@ -36,7 +54,14 @@ public class TeamController {
     }
 
     @PutMapping("/addCoach")
-    private ResponseEntity<?> addCoach(@RequestBody UserTeam request) {
+    private ResponseEntity<?> addCoach(
+            @RequestHeader(name = "auth-token") UUID authToken,
+            @RequestBody UserTeam request) {
+        boolean accessible = authService.checkAuthAndPermission(authToken, ROLE_ADMIN, ROLE_MANAGER);
+        if (!accessible) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         boolean added = teamService.addCoach(request.getTeamId(), request.getUserId());
         return added
                 ? new ResponseEntity<>(HttpStatus.OK)
@@ -44,7 +69,14 @@ public class TeamController {
     }
 
     @PutMapping("/addManager")
-    private ResponseEntity<?> addManager(@RequestBody UserTeam request) {
+    private ResponseEntity<?> addManager(
+            @RequestHeader(name = "auth-token") UUID authToken,
+            @RequestBody UserTeam request) {
+        boolean accessible = authService.checkAuthAndPermission(authToken, ROLE_ADMIN);
+        if (!accessible) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         boolean added = teamService.addManager(request.getTeamId(), request.getUserId());
         return added
                 ? new ResponseEntity<>(HttpStatus.OK)
@@ -52,7 +84,14 @@ public class TeamController {
     }
 
     @PutMapping("/addPlayer")
-    private ResponseEntity<?> addPlayer(@RequestBody UserTeam request) {
+    private ResponseEntity<?> addPlayer(
+            @RequestHeader(name = "auth-token") UUID authToken,
+            @RequestBody UserTeam request) {
+        boolean accessible = authService.checkAuthAndPermission(authToken, ROLE_ADMIN, ROLE_MANAGER, ROLE_COACH);
+        if (!accessible) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         boolean added = teamService.addPlayer(request.getTeamId(), request.getUserId());
         return added
                 ? new ResponseEntity<>(HttpStatus.OK)
@@ -60,7 +99,14 @@ public class TeamController {
     }
 
     @DeleteMapping("/deleteCoach")
-    private ResponseEntity<?> deleteCoach(@RequestBody UserTeam request) {
+    private ResponseEntity<?> deleteCoach(
+            @RequestHeader(name = "auth-token") UUID authToken,
+            @RequestBody UserTeam request) {
+        boolean accessible = authService.checkAuthAndPermission(authToken, ROLE_ADMIN, ROLE_MANAGER);
+        if (!accessible) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         boolean deleted = teamService.deleteCoach(request.getTeamId(), request.getUserId());
         return deleted
                 ? new ResponseEntity<>(HttpStatus.OK)
@@ -68,7 +114,14 @@ public class TeamController {
     }
 
     @DeleteMapping("/deleteManager")
-    private ResponseEntity<?> deleteManager(@RequestBody UserTeam request) {
+    private ResponseEntity<?> deleteManager(
+            @RequestHeader(name = "auth-token") UUID authToken,
+            @RequestBody UserTeam request) {
+        boolean accessible = authService.checkAuthAndPermission(authToken, ROLE_ADMIN);
+        if (!accessible) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         boolean deleted = teamService.deleteManager(request.getTeamId(), request.getUserId());
         return deleted
                 ? new ResponseEntity<>(HttpStatus.OK)
@@ -76,7 +129,14 @@ public class TeamController {
     }
 
     @DeleteMapping("/deletePlayer")
-    private ResponseEntity<?> deletePlayer(@RequestBody UserTeam request) {
+    private ResponseEntity<?> deletePlayer(
+            @RequestHeader(name = "auth-token") UUID authToken,
+            @RequestBody UserTeam request) {
+        boolean accessible = authService.checkAuthAndPermission(authToken, ROLE_ADMIN, ROLE_MANAGER, ROLE_COACH);
+        if (!accessible) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         boolean deleted = teamService.deletePlayer(request.getTeamId(), request.getUserId());
         return deleted
                 ? new ResponseEntity<>(HttpStatus.OK)
