+++ src/main/java/com/tsu/projectX/controllers/UserController.java
@@ -13,6 +13,8 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 import java.util.UUID;
 
+import static com.tsu.projectX.config.AuthConfig.*;
+
 @RestController()
 @CrossOrigin
 @RequestMapping(path = "/users")
@@ -20,19 +22,28 @@ public class UserController {
 
     @Autowired
     private IUserService userService;
-
     @Autowired
     private IAuthService authService;
 
+    @Deprecated
     @PostMapping
-    public ResponseEntity<?> create(@RequestBody User user) {
+    public ResponseEntity<?> create(
+            @RequestHeader(name = "auth-token") UUID authToken,
+            @RequestBody User user) {
         return userService.create(user)
                 ? new ResponseEntity<>(HttpStatus.CREATED)
                 : new ResponseEntity<>(HttpStatus.BAD_REQUEST);
     }
 
     @GetMapping("{id}")
-    public ResponseEntity<UserResponseDto> get(@PathVariable(name = "id") UUID id) {
+    public ResponseEntity<UserResponseDto> get(
+            @RequestHeader(name = "auth-token") UUID authToken,
+            @PathVariable(name = "id") UUID id) {
+        boolean accessible = authService.checkAuthAndPermission(authToken);
+        if (!accessible) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         UserResponseDto user = userService.get(id);
         return user != null
                 ? new ResponseEntity<>(user, HttpStatus.OK)
@@ -40,7 +51,13 @@ public class UserController {
     }
 
     @GetMapping()
-    public ResponseEntity<List<UserResponseDto>> getAll() {
+    public ResponseEntity<List<UserResponseDto>> getAll(
+            @RequestHeader(name = "auth-token") UUID authToken) {
+        boolean accessible = authService.checkAuthAndPermission(authToken);
+        if (!accessible) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         List<UserResponseDto> users = userService.getAll();
         return users != null && !users.isEmpty()
                 ? new ResponseEntity<>(users, HttpStatus.OK)
@@ -48,7 +65,15 @@ public class UserController {
     }
 
     @PutMapping("{id}")
-    public ResponseEntity<?> update(@PathVariable(name = "id") UUID id, @RequestBody UserRequestDto userRequestDto) {
+    public ResponseEntity<?> update(
+            @RequestHeader(name = "auth-token") UUID authToken,
+            @PathVariable(name = "id") UUID id,
+            @RequestBody UserRequestDto userRequestDto) {
+        boolean accessible = authService.checkAuthAndPermission(authToken, ROLE_ADMIN);
+        if (!accessible) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         boolean updated = userService.update(id, userRequestDto);
         return updated
                 ? new ResponseEntity<>(HttpStatus.OK)
@@ -56,7 +81,14 @@ public class UserController {
     }
 
     @DeleteMapping("{id}")
-    public ResponseEntity<?> delete(@PathVariable(name = "id") UUID id) {
+    public ResponseEntity<?> delete(
+            @RequestHeader(name = "auth-token") UUID authToken,
+            @PathVariable(name = "id") UUID id) {
+        boolean accessible = authService.checkAuthAndPermission(authToken, ROLE_ADMIN);
+        if (!accessible) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         boolean deleted = userService.delete(id);
         return deleted
                 ? new ResponseEntity<>(HttpStatus.OK)
